Backed out refactoring
